Make objectType an attribute on SolrDocument
